+++ common/foundation/src/form-association/submit-on-enter-key.ts
@@ -1,7 +1,7 @@
 import { requestSubmit } from '../form-association';
 import { getFormByIdOrClosest } from './common';
 
-export function submitOnEnter(element: HTMLInputElement) {
+export function submitOnEnter(element: HTMLInputElement): void {
 	element.addEventListener('keydown', function (
 		this: HTMLInputElement,
 		event: KeyboardEvent
